编译器优化，增加大数组读入功能,不再引发stack overflow的错误

+++ number.c
@@ -11,7 +11,7 @@ Object newNumber(double v){
 
 void numberFormat(char* des, Object num){
 	double v = GET_NUM(num);
-	if (fabs(v) - fabs((long) v) < 0.000000001) {
+	if (fabs(v) - fabs((double)(long) v) < 0.000000001) {
 		sprintf(des, "%ld", (long) v);
 	} else {
 		sprintf(des, "%g", v);
